AA-147 AA-134 N-03: remove aggregator from IAccount (#216)
* validateUserOp no longer receives an "aggregator" parameter.
* instead, validateUserOp returns "validationData" which includes aggregator, validAfter, validFrom
* helpers packValidationData, unpackValidationData can be used to pack and unpack this structure.
* removed IAggregatedAccount interface, as getAggregator is no longer needed.
* moved all validateSignatures calls to the beginning of the handleAggregatedOps, before any mutable calls of wallets or paymasters

+++ contracts/interfaces/IEntryPoint.sol
@@ -98,7 +98,7 @@ interface IEntryPoint is IStakeManager {
     /**
      * return value of simulateHandleOp
      */
-    error ExecutionResult(uint256 preOpGas, uint256 paid, uint64 validAfter, uint64 validUntil, bool targetSuccess, bytes targetResult);
+    error ExecutionResult(uint256 preOpGas, uint256 paid, uint48 validAfter, uint48 validUntil, bool targetSuccess, bytes targetResult);
 
     //UserOps handled, per aggregator
     struct UserOpsPerAggregator {
@@ -157,8 +157,8 @@ interface IEntryPoint is IStakeManager {
         uint256 preOpGas;
         uint256 prefund;
         bool sigFailed;
-        uint64 validAfter;
-        uint64 validUntil;
+        uint48 validAfter;
+        uint48 validUntil;
         bytes paymasterContext;
     }
